Create Whistle Noise for Modified Ska Reggae, Jamaican Rhythm.rb
For some mysterious reason, the song does not play unless you run the program twice. I believe that this may be due to some issue with the sync and cue. If anyone could check this out, solve the problem and let me know, then that would be awesome. Share with me your thoughts and I hope that you enjoy the track. Have a nice day!